added some tests for url_to_destination, and moved to elasticapm.utils
