+++ elasticapm/instrumentation/packages/asyncio/aiohttp_client.py
@@ -31,9 +31,8 @@
 from elasticapm import async_capture_span
 from elasticapm.conf import constants
 from elasticapm.instrumentation.packages.asyncio.base import AsyncAbstractInstrumentedModule
-from elasticapm.instrumentation.packages.http_utils import url_to_destination
 from elasticapm.traces import DroppedSpan, execution_context
-from elasticapm.utils import get_host_from_url, sanitize_url
+from elasticapm.utils import get_host_from_url, sanitize_url, url_to_destination
 from elasticapm.utils.disttracing import TracingOptions
