+++ elasticapm/instrumentation/packages/urllib3.py
@@ -30,9 +30,8 @@
 
 from elasticapm.conf import constants
 from elasticapm.instrumentation.packages.base import AbstractInstrumentedModule
-from elasticapm.instrumentation.packages.http_utils import url_to_destination
 from elasticapm.traces import DroppedSpan, capture_span, execution_context
-from elasticapm.utils import default_ports
+from elasticapm.utils import default_ports, url_to_destination
 from elasticapm.utils.disttracing import TracingOptions
